remove v0/showcase endpoints, tables, models, services
