+++ services/traction/api/db/models/__init__.py
@@ -2,12 +2,6 @@
 
 from api.db.models.base import BaseTable, Timeline  # noqa: F401
 from api.db.models.tenant import Tenant  # noqa: F401
-from api.db.models.tenant_webhook import TenantWebhook  # noqa: F401
-from api.db.models.tenant_webhook_msg import TenantWebhookMsg  # noqa: F401
-from api.db.models.issue_credential import IssueCredential  # noqa: F401
-from api.db.models.present_credential import PresentCredential  # noqa: F401
-from api.db.models.tenant_issuer import TenantIssuer  # noqa: F401
-from api.db.models.tenant_workflow import TenantWorkflow  # noqa: F401
 
 from api.db.models.v1.contact import Contact  # noqa: F401
 from api.db.models.v1.connection_invitation import ConnectionInvitation  # noqa: F401
@@ -31,8 +25,6 @@ from api.db.models.v1.tenant_configuration import (
 __all__ = [
     "BaseTable",
     "Tenant",
-    "TenantWebhook",
-    "TenantWebhookMsg",
     "Contact",
     "ConnectionInvitation",
     "SchemaTemplate",
