--- services/traction/api/db/repositories/tenant_issuers.py
@@ -1,83 +1,0 @@
-from typing import Type
-from uuid import UUID
-from sqlalchemy import select
-
-from api.db.errors import DoesNotExist
-from api.db.models.tenant_issuer import (
-    TenantIssuerCreate,
-    TenantIssuerUpdate,
-    TenantIssuerRead,
-    TenantIssuer,
-)
-from api.db.repositories.base import BaseRepository
-
-
-class TenantIssuersRepository(
-    BaseRepository[
-        TenantIssuerCreate, TenantIssuerUpdate, TenantIssuerRead, TenantIssuer
-    ]
-):
-    @property
-    def _in_schema(self) -> Type[TenantIssuerCreate]:
-        return TenantIssuerCreate
-
-    @property
-    def _upd_schema(self) -> Type[TenantIssuerUpdate]:
-        return TenantIssuerUpdate
-
-    @property
-    def _schema(self) -> Type[TenantIssuerRead]:
-        return TenantIssuerRead
-
-    @property
-    def _table(self) -> Type[TenantIssuer]:
-        return TenantIssuer
-
-    async def get_by_tenant_id(self, tenant_id: UUID) -> Type[TenantIssuerRead]:
-        q = select(self._table).where(self._table.tenant_id == tenant_id)
-        result = await self._db_session.execute(q)
-        tenant_issuer = result.scalar_one_or_none()
-        if not tenant_issuer:
-            raise DoesNotExist(
-                f"{self._table.__name__}<tenant_id:{tenant_id}> does not exist"
-            )
-        return self._schema.from_orm(tenant_issuer)
-
-    async def get_by_wallet_id(self, wallet_id: UUID) -> Type[TenantIssuerRead]:
-        q = select(self._table).where(self._table.wallet_id == wallet_id)
-        result = await self._db_session.execute(q)
-        tenant_issuer = result.scalar_one_or_none()
-        if not tenant_issuer:
-            raise DoesNotExist(
-                f"{self._table.__name__}<wallet_id:{wallet_id}> does not exist"
-            )
-        return self._schema.from_orm(tenant_issuer)
-
-    async def get_by_workflow_id(self, workflow_id: UUID) -> Type[TenantIssuerRead]:
-        q = select(self._table).where(self._table.workflow_id == workflow_id)
-        result = await self._db_session.execute(q)
-        tenant_issuer = result.scalar_one_or_none()
-        if not tenant_issuer:
-            raise DoesNotExist(
-                f"{self._table.__name__}<workflow_id:{workflow_id}> does not exist"
-            )
-        return self._schema.from_orm(tenant_issuer)
-
-    async def get_by_wallet_and_endorser_connection_id(
-        self, wallet_id: UUID, endorser_connection_id: UUID
-    ) -> Type[TenantIssuerRead]:
-        q = (
-            select(self._table)
-            .where(self._table.wallet_id == wallet_id)
-            .where(self._table.endorser_connection_id == endorser_connection_id)
-        )
-        result = await self._db_session.execute(q)
-        tenant_issuer = result.scalar_one_or_none()
-        if not tenant_issuer:
-            raise DoesNotExist(
-                (
-                    f"{self._table.__name__}<wallet_id:{wallet_id}>"
-                    f"<endorser_connection_id:{endorser_connection_id}> does not exist"
-                )
-            )
-        return self._schema.from_orm(tenant_issuer)
