--- services/traction/api/endpoints/models/tenant_issuer.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class PublicDIDStateType(str, Enum):
-    private = "private"
-    requested = "requested"
-    endorsed = "endorsed"
-    published = "published"
-    public = "public"
