--- services/traction/api/endpoints/models/tenant_schema.py
@@ -1,7 +1,0 @@
-from api.db.models.base import BaseSchema
-
-
-class TenantSchemaRequest(BaseSchema):
-    schema_name: str
-    schema_version: str
-    attributes: list[str]
