--- services/traction/api/endpoints/models/tenant_workflow.py
@@ -1,16 +1,0 @@
-from enum import Enum
-
-
-class TenantWorkflowTypeType(str, Enum):
-    connection = "api.services.ConnectionWorkflow"
-    issuer = "api.services.IssuerWorkflow"
-    schema = "api.services.SchemaWorkflow"
-    issue_cred = "api.services.IssueCredentialWorkflow"
-    present_cred = "api.services.PresentCredentialWorkflow"
-
-
-class TenantWorkflowStateType(str, Enum):
-    pending = "pending"
-    in_progress = "in_progress"
-    completed = "completed"
-    error = "error"
