+++ services/traction/api/protocols/v1/holder/holder_presentation_protocol.py
@@ -1,7 +1,7 @@
 from api.core.profile import Profile
 from api.db.models.v1.holder import HolderPresentation
 from api.db.session import async_session
-from api.endpoints.models.credentials import PresentationRoleType
+from api.endpoints.models.v1.enumerated import PresentationRoleType
 from api.endpoints.models.v1.errors import NotFoundError
 from api.endpoints.models.v1.holder import HolderPresentationStatusType
 from api.endpoints.models.v1.verifier import AcapyPresentProofStateType
